+++ mesh-cni/src/kubernetes/service.rs
@@ -263,7 +263,7 @@ fn generate_endpoint_events<T: KubeStore<EndpointSlice>>(
                         events.push(EndpointEvent::Delete(ServiceKeyV4 {
                             ip: ipv4_addr.into(),
                             port: np.port,
-                            protocol: np.protocol,
+                            protocol: np.protocol as u8,
                         }));
                     }
                     // TODO: Add Ipv6
@@ -297,7 +297,7 @@ fn generate_endpoint_events<T: KubeStore<EndpointSlice>>(
                     let service_destination = ServiceKeyV4 {
                         ip: ipv4_addr.into(),
                         port: np.port,
-                        protocol: np.protocol,
+                        protocol: np.protocol as u8,
                     };
                     let ready_destinations = destinations_map.remove(&np.name).unwrap_or_default();
                     events.push(EndpointEvent::Update(ServiceIdentity {
@@ -354,7 +354,7 @@ fn destinations_from_ep_slice(slice: &EndpointSlice) -> BTreeMap<String, Vec<Ser
             );
             continue;
         };
-        let protocol = KubeProtocol::try_from(port_proto_name.1.as_str()).unwrap_or_default();
+        let protocol = KubeProtocol::try_from(port_proto_name.1.as_str()).unwrap_or_default() as u8;
         let destinations: Vec<ServiceKeyV4> = addrs.iter().filter_map(|addr| match addr.parse::<IpAddr>(){
             Ok(ip) => {
                 match ip{
@@ -478,12 +478,12 @@ mod test {
                 service_destination: ServiceKeyV4 {
                     ip: Ipv4Addr::new(10, 96, 0, 25).into(),
                     port: 8080,
-                    protocol: KubeProtocol::Tcp,
+                    protocol: KubeProtocol::Tcp as u8,
                 },
                 ready_destinations: vec![ServiceKeyV4 {
                     ip: Ipv4Addr::new(192, 168, 1, 1).into(),
                     port: 80,
-                    protocol: KubeProtocol::Tcp,
+                    protocol: KubeProtocol::Tcp as u8,
                 }],
                 cluster_id: 1,
             })]
@@ -495,7 +495,7 @@ mod test {
             vec![EndpointEvent::Delete(ServiceKeyV4 {
                 ip: Ipv4Addr::new(10, 96, 0, 25).into(),
                 port: 8080,
-                protocol: KubeProtocol::Tcp,
+                protocol: KubeProtocol::Tcp as u8,
             })]
         );
     }
@@ -527,7 +527,7 @@ mod test {
             vec![ServiceKeyV4 {
                 ip: Ipv4Addr::new(192, 168, 1, 1).into(),
                 port: 80,
-                protocol: KubeProtocol::Tcp,
+                protocol: KubeProtocol::Tcp as u8,
             }],
         );
         assert_eq!(destinations_from_ep_slice(&slice), map);
